% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_blog_followed_by}
\alias{get_blog_followed_by}
\title{Check If Followed By Blog
This method can be used to check if one of your blogs is followed by another blog.}
\usage{
get_blog_followed_by(blog, query, app_credentials = NULL)
}
\arguments{
\item{blog}{name of the blog}

\item{query}{string. The name of the blog that may be following your blog}

\item{app_credentials}{a named list containing the consumer key and consumer secret. If NULL, attempts to load from an env variable}
}
\value{
logical
}
\description{
Check If Followed By Blog
This method can be used to check if one of your blogs is followed by another blog.
}
\examples{
\dontrun{
# replace "your-blog-name" with your Tumblr username
get_blog_followed_by(blog = "your-blog-name", query = "blog-to-check")
}
}
