% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_blog_following}
\alias{get_blog_following}
\title{Retrieve following}
\usage{
get_blog_following(
  blog,
  limit = 20,
  offset = 0,
  n = limit,
  app_credentials = NULL,
  ...
)
}
\arguments{
\item{blog}{name of the blog}

\item{limit}{The number of results to return per request: 1-20}

\item{offset}{post index to start at}

\item{n}{maximum number of posts to retrieve across pages (default \code{limit}, i.e. a single page). Use \code{Inf} to retrieve all posts of the blog.}

\item{app_credentials}{a named list containing the consumer key and consumer secret. If NULL, attempts to load from an env variable}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2}}
}
\value{
a tibble of blogs
}
\description{
This method can be used to retrieve the publicly exposed list of blogs that a blog follows, in order from most recently-followed to first.
\strong{Only works with your own account}
}
\examples{
\dontrun{
# replace "your-blog-name" with your Tumblr username
get_blog_following(blog = "your-blog-name")
}
}
