% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_blog_likes}
\alias{get_blog_likes}
\title{Retrieve Blog's Likes}
\usage{
get_blog_likes(blog, limit = 20, before, after, offset, api_key = NULL)
}
\arguments{
\item{blog}{name of the blog}

\item{limit}{The number of results to return: 1–20}

\item{before}{Retrieve posts liked before the specified timestamp}

\item{after}{Retrieve posts liked after the specified timestamp}

\item{offset}{Liked post number to start at}

\item{api_key}{}
}
\value{
tibble of liked blog posts
}
\description{
This method can be used to retrieve the publicly exposed likes from a blog.
}
\details{
You can only provide either before, after, or offset. If you provide more than one of these options together you will get an error.
You can still use limit with any of those three options to limit your result set.
When using the offset parameter the maximum limit on the offset is 1000. If you would like to get more results than that use either before or after.
}
