% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_dashboard}
\alias{get_dashboard}
\title{Retrieve a User's Dashboard}
\usage{
get_dashboard(limit = 20, offset = 0, n = limit, app_credentials = NULL, ...)
}
\arguments{
\item{limit}{The number of results to return per request: 1-20}

\item{offset}{post index to start at}

\item{n}{maximum number of posts to retrieve across pages (default \code{limit}, i.e. a single page). Use \code{Inf} to retrieve all posts of the blog.}

\item{app_credentials}{a named list containing the consumer key and consumer secret. If NULL, attempts to load from an env variable}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2}}
}
\value{
a tibble of posts
}
\description{
Returns the posts in the authenticating user's dashboard.
Requires OAuth authentication (see \code{\link[=rtumblr_auth]{rtumblr_auth()}}).
}
\examples{
\dontrun{
get_dashboard()
}
}
