% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_post}
\alias{get_post}
\title{Retrieve a Single Post}
\usage{
get_post(blog, post_id, api_key = NULL)
}
\arguments{
\item{blog}{name of the blog}

\item{post_id}{the id of the post to retrieve}

\item{api_key}{app consumer key. If NULL, attempts to load from the environment variable RTUMBLR_TOKEN}
}
\value{
a tibble with the post
}
\description{
Fetch a specific post by its id (new post format).
}
\examples{
\dontrun{
get_post("blog-name", post_id = "1234567890")
}
}
