% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_posts_tag}
\alias{get_posts_tag}
\title{Get Posts with Tag}
\usage{
get_posts_tag(tag, before, limit = 20, n = limit, api_key = NULL, ...)
}
\arguments{
\item{tag}{tag to search for}

\item{before}{the timestamp of when you'd like to see posts before}

\item{limit}{number of results to return per request: 1-20}

\item{n}{maximum number of posts to retrieve across pages (default \code{limit}, i.e. a single page). Use \code{Inf} to keep paging.}

\item{api_key}{app consumer key. If NULL, attempts to load from the environment variable RTUMBLR_TOKEN}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2}}
}
\value{
a list of tibbles of blog posts by format of posts
}
\description{
Get Posts with Tag
}
\details{
This function uses the legacy post format since it appears to not support the new post format.
The \verb{/tagged} endpoint does not support \code{offset}; pagination is done with the \code{before}
timestamp, which is handled automatically when \code{n} is larger than \code{limit}.
}
\examples{
\dontrun{
get_posts_tag(tag="meme")
# retrieve up to 100 posts across pages
get_posts_tag(tag="meme", n = 100)
}
}
