% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_tagged}
\alias{get_tagged}
\title{Get Posts with Tag}
\usage{
get_tagged(tag, limit = 20, api_key = NULL, ...)
}
\arguments{
\item{tag}{tag to search for}

\item{limit}{The number of results to return: 1–50}

\item{api_key}{app consumer key. If NULL, attempts to load from an env variable}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2}}
}
\value{
a tibble of blog posts
}
\description{
Get Posts with Tag
}
\details{
This function uses the legacy post format since it appears to not suppord the new post format
}
