% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_info}
\alias{get_user_info}
\title{Retrieve a User's Information}
\usage{
get_user_info(app_credentials = NULL)
}
\arguments{
\item{app_credentials}{a named list containing the consumer key and consumer secret. If NULL, attempts to load from an env variable}
}
\value{
a tibble with the user's information; the user's blogs are stored in the \code{blogs} list column
}
\description{
Returns information about the authenticating user, including which blogs they own.
Requires OAuth authentication (see \code{\link[=rtumblr_auth]{rtumblr_auth()}}).
}
\examples{
\dontrun{
get_user_info()
}
}
