% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_limits}
\alias{get_user_limits}
\title{Retrieve a User's Limits}
\usage{
get_user_limits(app_credentials = NULL)
}
\arguments{
\item{app_credentials}{a named list containing the consumer key and consumer secret. If NULL, attempts to load from an env variable}
}
\value{
a list with the user's limits as returned by the API
}
\description{
Returns the authenticating user's rate/usage limits.
Requires OAuth authentication (see \code{\link[=rtumblr_auth]{rtumblr_auth()}}).
}
\examples{
\dontrun{
get_user_limits()
}
}
