% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_blog_info}
\alias{get_blog_info}
\title{Retrieve Blog Info}
\usage{
get_blog_info(blog, api_key = NULL)
}
\arguments{
\item{blog}{name of the blog}

\item{api_key}{app consumer key. If NULL, attempts to load from the environment variable RTUMBLR_TOKEN}
}
\value{
tibble of information about  blog
}
\description{
This method returns general information about the blog, such as the title, number of posts, and other high-level data.
}
\examples{
\dontrun{
get_blog_info("schochastics")
}
}
