% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_blog_posts}
\alias{get_blog_posts}
\title{Retrieve Published Posts}
\usage{
get_blog_posts(blog, limit = 50, offset = 0, api_key = NULL, ...)
}
\arguments{
\item{blog}{name of the blog}

\item{limit}{The number of results to return: 1–50}

\item{offset}{post index to start at}

\item{api_key}{app consumer key}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2#posts--retrieve-published-posts}}
}
\value{
a tibble of blog posts
}
\description{
Retrieve Published Posts
}
