% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_blog_posts}
\alias{get_blog_posts}
\title{Retrieve Published Posts}
\usage{
get_blog_posts(blog, limit = 50, offset = 0, api_key = NULL, ...)
}
\arguments{
\item{blog}{name of the blog}

\item{limit}{The number of results to return: 1–50}

\item{offset}{post index to start at}

\item{api_key}{app consumer key. If NULL, attempts to load from the environment variable RTUMBLR_TOKEN}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2}}
}
\value{
a tibble of blog posts
}
\description{
Retrieve Published Posts
}
\details{
this function uses the new post format (npf: \url{https://www.tumblr.com/docs/npf})
}
\examples{
\dontrun{
# replace "blog-name" with a Tumblr username
get_blog_posts(blog = "blog-name")
}
}
