% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogs.R
\name{get_posts_tag}
\alias{get_posts_tag}
\title{Get Posts with Tag}
\usage{
get_posts_tag(tag, before, limit = 20, api_key = NULL, ...)
}
\arguments{
\item{tag}{tag to search for}

\item{before}{the timestamp of when you'd like to see posts before}

\item{limit}{The number of results to return: 1–50}

\item{api_key}{app consumer key. If NULL, attempts to load from the environment variable RTUMBLR_TOKEN}

\item{...}{further parameters as described here: \url{https://www.tumblr.com/docs/en/api/v2}}
}
\value{
a list of tibbles of blog posts by format of posts
}
\description{
Get Posts with Tag
}
\details{
This function uses the legacy post format since it appears to not support the new post format
}
\examples{
\dontrun{
get_posts_tag(tag="meme")
}
}
